\name{extract.forecast.cov}
\alias{extract.forecast.cov}
\alias{extract.forecast.cov.estimators.wrt.data}
\alias{extract.forecast.cov.estimators.from.model}

\title{Extract Forecast Covariance}
\description{extract.forecast.cov}
\usage{
    extract.forecast.cov(e, n)
}
\arguments {
    \item{e}{A forecast.cov object.}
    \item{n}{A vector on integers.}
}
\value{A forecast.cov object.}
\details{
    Select a subset of models and their forecast covariances from a
    larger object.
}
\seealso{
    \code{\link{forecast.cov}}
}

%\keyword{DSE}
\keyword{ts}

