\name{input.data.TSPADIdata}
\alias{input.data.TSPADIdata}
\alias{input.periods.TSPADIdata}
\alias{input.series.names.feather.forecasts}
\alias{input.series.names.forecast}
\alias{input.series.names.monte.carlo.simulations}
\alias{output.data.TSPADIdata}
\alias{output.periods.TSPADIdata}
\alias{output.series.names.feather.forecasts}
\alias{output.series.names.forecast}
\alias{output.series.names.monte.carlo.simulations}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
	input.data(x, ...)
	output.data(x, ...)
	input.periods(x, ...)
	output.periods(x, ...)
	input.series.names(x, ...)
	output.series.names(x, ...)
}

\keyword{DSE}
\keyword{ts}

