\name{input.series.names.forecast}
\alias{input.series.names.forecast}
\alias{input.series.names.featherForecasts}
\alias{input.series.names.MonteCarloSimulations}

\alias{output.series.names.forecast}
\alias{output.series.names.featherForecasts}
\alias{output.series.names.MonteCarloSimulations}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
    \method{input.series.names}{forecast}(obj)
    \method{input.series.names}{featherForecasts}(obj)
    \method{input.series.names}{MonteCarloSimulations}(obj)

    \method{output.series.names}{forecast}(obj)
    \method{output.series.names}{featherForecasts}(obj)
    \method{output.series.names}{MonteCarloSimulations}(obj)
}
\arguments{
    \item{obj}{an object from which to extract the names of the input or output
      series.}
}

%\keyword{DSE}
\keyword{ts}

