\name{is.forecast.cov.estimators.wrt.data.subsets}
\alias{is.forecast.cov.estimators.wrt.data.subsets}
\title{Check Inheritance}
\description{Check inheritance.}
\usage{
    is.forecast.cov.estimators.wrt.data.subsets(obj)
}
\arguments {
    \item{obj}{Any object.}
}
\value{logical}
\details{
    This tests if an object isnherits 
    from \code{forecast.cov.estimators.wrt.data.subsets}. This type of object
    code be generated in different ways but the only current example
    is \code{mine.strip}.
}
\seealso{
    \code{\link{mine.strip}}
}
\
%\keyword{DSE}
\keyword{ts}

