\name{is.forecastCov.estimatorsWRTdata.subsets}
\alias{is.forecastCov.estimatorsWRTdata.subsets}
\title{Check Inheritance}
\description{Check inheritance.}
\usage{
    is.forecastCov.estimatorsWRTdata.subsets(obj)
}
\arguments{
    \item{obj}{Any object.}
}
\value{logical}
\details{
    This tests if an object isnherits 
    from \code{forecastCov.estimatorsWRTdata.subsets}. This type of object
    code be generated in different ways but the only current example
    is \code{mine.strip}.
}
\seealso{
    \code{\link{mine.strip}}
}
\
%\keyword{DSE}
\keyword{ts}

