\name{plot.mine.stepwise}
\alias{plot.mine.stepwise}
\title{Plot Mine Stepwise Object}
\description{plot.mine.stepwise}
\usage{
    plot.mine.stepwise(x)
}
\arguments {
    \item{x}{Object returned by mine.stepwise.}
}
\value{None}
\sideffects{A plot}
\seealso{
    \code{\link{mine.stepwise}}
}
%\keyword{DSE}
\keyword{ts}
