\name{plot.mineStepwise}
\alias{plot.mineStepwise}
\title{Plot Mine Stepwise Object}
\description{plot.mineStepwise}
\usage{
    \method{plot}{mineStepwise}(x, ...)
}
\arguments{
    \item{x}{Object returned by mineStepwise.}
    \item{...}{(further arguments, currently disregarded).}
}
\value{None}
\sideffects{A plot}
\seealso{
    \code{\link{mineStepwise}}
}
%\keyword{DSE}
\keyword{ts}
