\name{roots.estimated.models}
\alias{roots.estimated.models}
\alias{roots.forecast.cov.estimators.wrt.true}
\alias{roots.parms.ee}
\alias{roots.roots.ee}
\alias{roots.TSestModel.ee}
\alias{roots.TSmodel.ee}

\title{Roots Specific Methods}
\description{See the generic function description.}
\usage{
    roots(obj)
}

\seealso{ \code{\link{stability}} }

%\keyword{DSE}
\keyword{ts}

