\name{seriesNamesInput.forecast}
\alias{seriesNamesInput.forecast}
\alias{seriesNamesInput.featherForecasts}
\alias{seriesNamesInput.MonteCarloSimulations}

\alias{seriesNamesOutput.forecast}
\alias{seriesNamesOutput.featherForecasts}
\alias{seriesNamesOutput.MonteCarloSimulations}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
    \method{seriesNamesInput}{forecast}(obj)
    \method{seriesNamesInput}{featherForecasts}(obj)
    \method{seriesNamesInput}{MonteCarloSimulations}(obj)

    \method{seriesNamesOutput}{forecast}(obj)
    \method{seriesNamesOutput}{featherForecasts}(obj)
    \method{seriesNamesOutput}{MonteCarloSimulations}(obj)
}
\arguments{
    \item{obj}{an object from which to extract the names of the input or output
      series.}
}

%\keyword{DSE}
\keyword{ts}

