\name{summary.TSPADIdata}
\alias{summary.TSestModel.ee}
\alias{summary.TSmodel.ee}
\alias{summary.estimated.models}
\alias{summary.estimation.evaluation}
\alias{summary.forecast.cov}
\alias{summary.forecast.cov.estimators.wrt.data.subsets}
\alias{summary.forecast.cov.estimators.wrt.true}
\alias{summary.monte.carlo.simulations}
\alias{summary.parms.ee}
\alias{summary.roots.ee}

\alias{print.summary.TSestModel.ee}
\alias{print.summary.TSmodel.ee}
\alias{print.summary.estimated.models}
\alias{print.summary.estimation.evaluation}
\alias{print.summary.forecast.cov}
\alias{print.summary.forecast.cov.estimators.wrt.data.subsets}
\alias{print.summary.forecast.cov.estimators.wrt.true}
\alias{print.summary.monte.carlo.simulations}
\alias{print.summary.parms.ee}
\alias{print.summary.roots.ee}

\title{Summary Specific Methods}
\description{See the generic function description.}
\usage{
	summary(x, ...)
}

\seealso{
	\code{\link{summary}}
	\code{\link{print}}
}

\keyword{DSE}
\keyword{ts}

