\name{TSdata.forecast.cov}
\alias{TSdata.forecast.cov}
\alias{TSestModel.parms.ee}
\alias{TSmodel.forecast.cov}
\alias{TSmodel.parms.ee}

\title{TS Extractor Specific Methods}
\description{See the generic function description.}
\usage{
	TSdata(x, ...)
	TSestModel(x, ...)
	TSmodel(x, ...)
}

\seealso{ 
	\code{\link{TSdata}} 
	\code{\link{TSestModel}} 
	\code{\link{TSmodel}} 
}

\keyword{DSE}
\keyword{ts}

