
\name{dse2-package}
\alias{dse2-package}

\docType{package}

\title{Transition Package from dse Bundle to EvalEst Package.}

\description{Package to add transition from dse Bundle to EvalEst package.
}

\section{Usage}{
   library("dse2") 
}

\details{
The only purpose of this package is to prevent breaking older code that loads 
package dse2 from the old dse bundle. This package dse2 simple requires package
EvalEst.
}

\author{
Paul Gilbert <pgilbert@bank-banque-canada.ca>

Maintainer: Paul Gilbert <pgilbert@bank-banque-canada.ca>  
}
