\name{input.series.names.forecast}
\alias{input.series.names.feather.forecasts}
\alias{input.series.names.forecast}
\alias{input.series.names.monte.carlo.simulations}
\alias{output.series.names.feather.forecasts}
\alias{output.series.names.forecast}
\alias{output.series.names.monte.carlo.simulations}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
    input.series.names(x, ...)
    output.series.names(x, ...)
}
%\keyword{DSE}
\keyword{ts}

