\name{test.equal.estimated.models}
\alias{test.equal.estimated.models}
\alias{test.equal.estimation.evaluation}
\alias{test.equal.forecast}
\alias{test.equal.forecast.cov}
\alias{test.equal.horizon.forecasts}
\alias{test.equal.monte.carlo.simulations}

\title{Specific Methods for Testing Equality}
\description{See the generic function description.}
\usage{
	test.equal(obj1, obj2)
}

\seealso{ \code{\link{test.equal}} }

\keyword{DSE}
\keyword{ts}

